md.Parse should be accept config, so change interface
