+++ cmd/export.go
@@ -42,7 +42,7 @@ var exportCmd = &cobra.Command{
 		ctx := cmd.Context()
 		if len(args) > 0 {
 			f := args[0]
-			markdownData, err := md.ParseFile(f)
+			markdownData, err := md.ParseFile(f, nil)
 			if err != nil {
 				// Deprecated
 				presentationID = f
