+++ md/md_test.go
@@ -8,8 +8,6 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/google/go-cmp/cmp"
-	"github.com/k1LoW/deck/config"
 	"github.com/tenntenn/golden"
 )
 
@@ -52,7 +50,7 @@ func TestParse(t *testing.T) {
 			if err != nil {
 				t.Fatal(err)
 			}
-			md, err := Parse("../testdata", b)
+			md, err := Parse("../testdata", b, nil)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -66,175 +64,6 @@ func TestParse(t *testing.T) {
 			}
 			if diff := golden.Diff(t, "", tt.in, got); diff != "" {
 				t.Error(diff)
-			}
-		})
-	}
-}
-
-func TestApplyConfig(t *testing.T) {
-	tests := []struct {
-		name               string
-		initialFrontmatter *Frontmatter
-		config             *config.Config
-		want               *Frontmatter
-	}{
-		{
-			name: "Apply config breaks when frontmatter breaks is not set",
-			initialFrontmatter: &Frontmatter{
-				Breaks: nil, // not set
-			},
-			config: &config.Config{
-				Breaks: boolPtr(true),
-			},
-			want: &Frontmatter{
-				Breaks:   boolPtr(true),
-				Defaults: nil,
-			},
-		},
-		{
-			name: "Keep existing breaks value when already set",
-			initialFrontmatter: &Frontmatter{
-				Breaks: boolPtr(false), // already set
-			},
-			config: &config.Config{
-				Breaks: boolPtr(true),
-			},
-			want: &Frontmatter{
-				Breaks:   boolPtr(false), // keep existing value
-				Defaults: nil,
-			},
-		},
-		{
-			name:               "Apply breaks when frontmatter is nil",
-			initialFrontmatter: nil,
-			config: &config.Config{
-				Breaks: boolPtr(true),
-			},
-			want: &Frontmatter{
-				Breaks:   boolPtr(true),
-				Defaults: nil,
-			},
-		},
-		{
-			name: "Add config defaults when no existing defaults conditions",
-			initialFrontmatter: &Frontmatter{
-				Defaults: []DefaultCondition{}, // empty slice
-			},
-			config: &config.Config{
-				Defaults: []config.DefaultCondition{
-					{
-						If:     "page == 1",
-						Layout: "title",
-						Freeze: boolPtr(true),
-					},
-				},
-			},
-			want: &Frontmatter{
-				Breaks: nil,
-				Defaults: []DefaultCondition{
-					{
-						If:     "page == 1",
-						Layout: "title",
-						Freeze: boolPtr(true),
-					},
-				},
-			},
-		},
-		{
-			name: "Append config defaults when existing defaults conditions present",
-			initialFrontmatter: &Frontmatter{
-				Defaults: []DefaultCondition{
-					{
-						If:     "page == 2",
-						Layout: "content",
-						Skip:   boolPtr(true),
-					},
-				},
-			},
-			config: &config.Config{
-				Defaults: []config.DefaultCondition{
-					{
-						If:     "page == 1",
-						Layout: "title",
-						Freeze: boolPtr(true),
-					},
-					{
-						If:     "page == 3",
-						Layout: "end",
-						Ignore: boolPtr(true),
-					},
-				},
-			},
-			want: &Frontmatter{
-				Breaks: nil,
-				Defaults: []DefaultCondition{
-					{
-						If:     "page == 2",
-						Layout: "content",
-						Skip:   boolPtr(true),
-					},
-					{
-						If:     "page == 1",
-						Layout: "title",
-						Freeze: boolPtr(true),
-					},
-					{
-						If:     "page == 3",
-						Layout: "end",
-						Ignore: boolPtr(true),
-					},
-				},
-			},
-		},
-		{
-			name: "Apply config codeBlockToImageCommand when frontmatter codeBlockToImageCommand is not set",
-			initialFrontmatter: &Frontmatter{
-				CodeBlockToImageCommand: "", // not set
-			},
-			config: &config.Config{
-				CodeBlockToImageCommand: "go run testdata/txt2img/main.go",
-			},
-			want: &Frontmatter{
-				CodeBlockToImageCommand: "go run testdata/txt2img/main.go",
-				Defaults:                nil,
-			},
-		},
-		{
-			name: "Keep existing codeBlockToImageCommand value when already set",
-			initialFrontmatter: &Frontmatter{
-				CodeBlockToImageCommand: "go run testdata/txt2img/main.go", // already set
-			},
-			config: &config.Config{
-				CodeBlockToImageCommand: "go run other/command",
-			},
-			want: &Frontmatter{
-				CodeBlockToImageCommand: "go run testdata/txt2img/main.go", // keep existing value
-				Defaults:                nil,
-			},
-		},
-		{
-			name:               "Apply codeBlockToImageCommand when frontmatter is nil",
-			initialFrontmatter: nil,
-			config: &config.Config{
-				CodeBlockToImageCommand: "go run testdata/txt2img/main.go",
-			},
-			want: &Frontmatter{
-				CodeBlockToImageCommand: "go run testdata/txt2img/main.go",
-				Defaults:                nil,
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			md := &MD{
-				Frontmatter: tt.initialFrontmatter,
-			}
-
-			md.ApplyConfig(tt.config)
-
-			if diff := cmp.Diff(tt.want, md.Frontmatter); diff != "" {
-				t.Errorf("Frontmatter mismatch (-want +got):\n%s", diff)
 			}
 		})
 	}
@@ -441,6 +270,6 @@ ref: [deck repo](https://github.com/k1LoW/deck)
 # Title
 `))
 	f.Fuzz(func(t *testing.T, in []byte) {
-		_, _ = Parse(".", in)
+		_, _ = Parse(".", in, nil)
 	})
 }
